// src/types/votes.ts
// Загальні типи, які використовуються в кількох компонентах, винесені в цей окремий файл
// ==========================================================================================
// Стан додатка
// ==========================================================================================
// При кліку на одну з кнопок ("Good", "Neutral", "Bad") з компонента VoteOptions
// потрібно зберігати інформацію про вибір користувача та передавати її до інших компонентів,
// зокрема – до того, який відповідає за відображення статистики <VoteStats/>.
// ------------------------------------------------------------------------------------------
// Для цього потрібен стан votes, який зберігатиме кількість голосів
// ------------------------------------------------------------------------------------------

// Інтерфейс для стану
export interface Votes {
  good: number;
  neutral: number;
  bad: number;
}

// Перелік допустимих рядкових значень можливого вибору
export type VoteType = "good" | "neutral" | "bad";